board/freescale/common/imx: align u-boot-spl to 4 bytes
When enable DM for SPL binary, the DTB part of SPL may not 4 bytes aligned.
If u-boot-spl is not aligned, the offset of the DDR firmware is not 4
byte aligned when u-boot-spl-ddr.bin is created. This causes the ddr
firmware to not be loaded correctly at boot.

See imx-mkimage commit bba038d893046b44683182dba540f104dab80fe7 for
implementation in imx-mkimage.

Signed-off-by: Bram Vlerick <bram.vlerick@openpixelsystems.org>